Authentication and Authorization added

+++ src/main/java/com/sagar/project/airBnbApp/service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.sagar.project.airBnbApp.dto.HotelInfoDto;
 import com.sagar.project.airBnbApp.dto.RoomDto;
 import com.sagar.project.airBnbApp.entity.Hotel;
 import com.sagar.project.airBnbApp.entity.Room;
+import com.sagar.project.airBnbApp.entity.User;
 import com.sagar.project.airBnbApp.exception.ResourceNotFoundException;
+import com.sagar.project.airBnbApp.exception.UnAuthorisedException;
 import com.sagar.project.airBnbApp.repository.HotelRepository;
 import com.sagar.project.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,17 +28,7 @@ public class HotelServiceImpl implements HotelService{
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
-        private final RoomRepository roomRepository ;
-
-    @Override
-    public HotelDto createNewHotel(HotelDto hotelDto) {
-        log.info("Creating a new hotel with name: {}", hotelDto.getName());
-        Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
-        hotel.setActive(false);
-        hotel = hotelRepository.save(hotel);
-        log.info("Created a new hotel with ID: {}", hotelDto.getId());
-        return modelMapper.map(hotel, HotelDto.class);
-    }
+    private final RoomRepository roomRepository;
 
     @Override
     public List<HotelDto> getAllHotels() {
@@ -45,11 +38,31 @@ public class HotelServiceImpl implements HotelService{
     }
 
     @Override
+    public HotelDto createNewHotel(HotelDto hotelDto) {
+        log.info("Creating a new hotel with name: {}", hotelDto.getName());
+        Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
+        hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
+        hotel = hotelRepository.save(hotel);
+        log.info("Created a new hotel with ID: {}", hotelDto.getId());
+        return modelMapper.map(hotel, HotelDto.class);
+    }
+
+    @Override
     public HotelDto getHotelById(Long id) {
         log.info("Getting the hotel with ID: {}", id);
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -59,6 +72,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -72,12 +91,17 @@ public class HotelServiceImpl implements HotelService{
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
+
         for(Room room: hotel.getRooms()) {
-            inventoryService.deleteAllInvertory(room);
+            inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
         }
         hotelRepository.deleteById(id);
-
     }
 
     @Override
@@ -88,6 +112,12 @@ public class HotelServiceImpl implements HotelService{
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         hotel.setActive(true);
 
         // assuming only do it once
@@ -96,12 +126,20 @@ public class HotelServiceImpl implements HotelService{
         }
     }
 
+    //    public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
-        List<RoomDto> rooms = hotel.getRooms().stream().map(room -> modelMapper.map(room, RoomDto.class)).collect(Collectors.toList());
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        List<RoomDto> rooms = hotel.getRooms()
+                .stream()
+                .map((element) -> modelMapper.map(element, RoomDto.class))
+                .toList();
+
         return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
     }
+
+
 }
